Got extra compiler warnings enabled and changed code to conform to them. ~300 warnings!

+++ src/Camera.cpp
@@ -7,10 +7,10 @@
 
 #include "Camera.hpp"
 
-Ray Camera::rayForPixel(const int px, const int py) const
+Ray Camera::rayForPixel(const uint32_t px, const uint32_t py) const
 {
-	const float xOffset = (px + 0.5) * pixelSize;
-	const float yOffset = (py + 0.5) * pixelSize;
+	const float xOffset = (static_cast<float>(px) + 0.5f) * pixelSize;
+	const float yOffset = (static_cast<float>(py) + 0.5f) * pixelSize;
 
 	const float worldX = halfWidth - xOffset;
 	const float worldY = halfHeight - yOffset;
@@ -26,9 +26,9 @@ Canvas Camera::Render(World w) const
 {
 	Canvas image = Canvas(hSize, vSize);
 
-	for (int i = 0; i < vSize; i++)
+	for (uint32_t i = 0; i < vSize; i++)
 	{
-		for (int j = 0; j < hSize; j++)
+		for (uint32_t j = 0; j < hSize; j++)
 		{
 			const Ray r = rayForPixel(j, i);
 			const Color c = w.colorAt(r);
